Moving `Progress` to own module
